Add support for importing BIP39 mnemonic from textfile
The main change in this commit adds a --private-file argument to view-addresses and
sign-psbt that enables users to load the BIP39 mnemonic from a 1 line textfile. This
feature is desirable because it means that users have the option to generate the BIP39
mnemonic once (back it up on paper), and save it to a USB (such as the one that also
stores bitcoind, zbarcam, and qrencode) so that the private key data need never be
visible on the user's laptop display.

Other changes in this commit include the deletion of some legacy tests and the
addition/update of other tests.